+++ CHANGELOG.md
@@ -2,13 +2,21 @@
 
 ### [Unreleased][unreleased]
 
+#### Changed
+- Removed parameter database from constructor - BC break (use command select )
+- Return values of some commands
+- Unified info command for both drivers (inspired by predis)
+
 #### Added
-- timeout to redis connect
-- possibility to set drivers order
+- Timeout to redis connect
+- Possibility to set drivers order
+- RedisProxyException - common exception for all drivers
+- Predis commands
 
 ### [0.1.0] - 2016-04-17
-- First tagged version
-- possible drivers: \Redis, \Predis\Client
+
+#### Added
+- Wrapper for \Redis
 
 [unreleased]: https://github.com/lulco/redis-proxy/compare/0.1.0...HEAD
 [0.1.0]: https://github.com/lulco/redis-proxy/compare/0.0.0...0.1.0